server/requirements.txt
Django>=3.2,<4.0
channels
mysqlclient
djangorestframework
djangorestframework-simplejwt
python-socketio
eventlet
tkinter
```